+++ src/edcon/edrive/parameter_set.py
@@ -2,61 +2,9 @@
 """
 Contains ParameterSet class which is used to represent parameter sets of EDrives.
 """
-from dataclasses import dataclass
+from edcon.edrive.parameter import Parameter
 from edcon.edrive.parameter_mapping import ParameterMap
 from edcon.utils.logging import Logging
-
-
-@dataclass
-class Parameter:
-    """Class representing a parameter."""
-    axis: int
-    data_id: int
-    instance: int
-    subindex: int
-    value: bytes
-
-    @classmethod
-    def from_uid(cls, uid_id: str, value: bytes):
-        """Function to create a Parameter by providing a uid and value.
-
-        Parameters:
-            uid_id (str): uid of the parameter
-            value (bytes): value of the parameter
-
-        Returns:
-            Parameter object
-        """
-        axis, data_id, instance, subindex = uid_id.strip('P').split('.')
-        return cls(int(axis), int(data_id), int(instance), int(subindex), value)
-
-    def uid(self) -> str:
-        """Returns the uid of the parameter.
-
-        Returns:
-            str: uid
-        """
-        return f'{self.axis}.{self.data_id}.{self.instance}.{self.subindex}'
-
-    def is_null_terminator(self):
-        """Determines if the parameter is a null terminator.
-
-        Returns:
-            bool: True if the parameter is a null terminator
-        """
-        parameter_map = ParameterMap()
-        if not self.uid() in parameter_map:
-            return None
-
-        data_type = parameter_map[self.uid()].data_type
-        if 'STRING' in data_type:
-            last_index = int(data_type.strip('STRING()')) - 1
-            if self.subindex == last_index:
-                Logging.logger.debug(
-                    f'Parameter {self.uid()}.{self.subindex} is a null terminator')
-                return True
-        return False
-
 
 class ParameterSet:
     """Class representing a parameter set."""
@@ -71,9 +19,9 @@ class ParameterSet:
 
         for item in lines[start_idx:end_idx]:
             key, hex_value = item.decode().strip('P').split(';')
-            value = bytes.fromhex(hex_value.split('x')[1])[::-1]
+            value_raw = bytes.fromhex(hex_value.split('x')[1])[::-1]
 
-            self.parameters.append(Parameter.from_uid(key, value))
+            self.parameters.append(Parameter.from_uid_raw(key, value_raw))
 
         if strip_null_terminators:
             Logging.logger.info('Stripping null terminators from parameter set')
@@ -85,7 +33,26 @@ class ParameterSet:
     def __len__(self):
         return len(self.parameters)
 
+    def _is_parameter_null_terminator(self, parameter: Parameter):
+        """Determines if the parameter is a null terminator.
+
+        Returns:
+            bool: True if the parameter is a null terminator
+        """
+        parameter_map = ParameterMap()
+        if not parameter.uid() in parameter_map:
+            return None
+
+        data_type = parameter_map[parameter.uid()].data_type
+        if 'STRING' in data_type:
+            last_index = int(data_type.strip('STRING()')) - 1
+            if parameter.subindex == last_index:
+                Logging.logger.debug(
+                    f'Parameter {parameter.uid()} is a null terminator')
+                return True
+        return False
+
     def strip_null_terminators(self):
         """Removes all parameters from the parameter set that represent null terminators."""
-        self.parameters = filter(
-            lambda v: not v.is_null_terminator(), self.parameters)
+        self.parameters = list(filter(
+            lambda v: not self._is_parameter_null_terminator(v), self.parameters))